Various bugfixes, renamed scripts
