+++ bin/produse
@@ -1,7 +1,7 @@
 #! /usr/bin/env python
 
 import sys
-from ProDuSe import ProdusePipeline, trim, collapse, adapter_predict
+from ProDuSe import ProdusePipeline, Trim, Collapse, AdapterPredict
 
 def usage():
     """
@@ -35,13 +35,20 @@ def main(args):
     command = args[1]
 
     if command == "adapter_predict":
-         adapter_predict.main(sysStdin=args[1:])
+         AdapterPredict.main(sysStdin=args[2:])
     elif command == "run_produse":
         ProdusePipeline.main(sysStdin=args[2:])
     elif command == "trim":
-        trim.main(sysStdin=args[2:])
+        Trim.main(sysStdin=args[2:])
     elif command == "collapse":
-        collapse.main(sysStdin=args[2:])
+        Collapse.main(sysStdin=args[2:])
+    elif command == "clip":
+        print("Not yet implemented")
+    elif command == "call":
+        print("Coming soon")
+    else:
+        sys.stdout.write("Unrecognized Command: %s\n" % command)
+        usage()
 
 if __name__ == "__main__":
     main(sys.argv)
